+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Rules/Ledger.hs
@@ -23,11 +23,11 @@ import Cardano.Ledger.BaseTypes (ShelleyBase)
 import Cardano.Ledger.Keys (DSignable, Hash)
 import Cardano.Ledger.Shelley.Core
 import Cardano.Ledger.Shelley.LedgerState (
-  DPState (..),
+  CertState (..),
   DState (..),
   LedgerState (..),
   UTxOState (..),
-  obligationDPState,
+  obligationCertState,
  )
 import Cardano.Ledger.Shelley.Rules (
   DelegsEnv (..),
@@ -72,7 +72,7 @@ ledgerTransition ::
   , Embed (EraRule "UTXOW" era) (someLEDGER era)
   , Embed (EraRule "DELEGS" era) (someLEDGER era)
   , Environment (EraRule "DELEGS" era) ~ DelegsEnv era
-  , State (EraRule "DELEGS" era) ~ DPState (EraCrypto era)
+  , State (EraRule "DELEGS" era) ~ CertState era
   , Signal (EraRule "DELEGS" era) ~ Seq (DCert (EraCrypto era))
   , Environment (EraRule "UTXOW" era) ~ UtxoEnv era
   , State (EraRule "UTXOW" era) ~ UTxOState era
@@ -95,7 +95,7 @@ ledgerTransition = do
             )
       else pure dpstate
 
-  let DPState dstate _pstate = dpstate
+  let CertState dstate _pstate _ = dpstate
       genDelegs = dsGenDelegs dstate
 
   utxoSt' <-
@@ -117,7 +117,7 @@ instance
   , State (EraRule "UTXOW" era) ~ UTxOState era
   , Signal (EraRule "UTXOW" era) ~ AlonzoTx era
   , Environment (EraRule "DELEGS" era) ~ DelegsEnv era
-  , State (EraRule "DELEGS" era) ~ DPState (EraCrypto era)
+  , State (EraRule "DELEGS" era) ~ CertState era
   , Signal (EraRule "DELEGS" era) ~ Seq (DCert (EraCrypto era))
   , ProtVerAtMost era 8
   ) =>
@@ -139,7 +139,7 @@ instance
     [ PostCondition
         "Deposit pot must equal obligation (AlonzoLEDGER)"
         ( \(TRC (_, _, _))
-           (LedgerState utxoSt dpstate) -> obligationDPState dpstate == utxosDeposited utxoSt
+           (LedgerState utxoSt dpstate) -> obligationCertState dpstate == utxosDeposited utxoSt
         )
     ]