+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Translation.hs
@@ -15,6 +15,7 @@ import Cardano.Ledger.Alonzo.PParams ()
 import Cardano.Ledger.Alonzo.Tx (AlonzoTx (..), IsValid (..))
 import Cardano.Ledger.Alonzo.TxBody (AlonzoTxOut (..))
 import Cardano.Ledger.Binary (DecoderError)
+import Cardano.Ledger.CertState (PState (..), VState (..))
 import Cardano.Ledger.Core (upgradePParams, upgradePParamsUpdate)
 import qualified Cardano.Ledger.Core as Core
 import Cardano.Ledger.Crypto (Crypto)
@@ -25,6 +26,8 @@ import Cardano.Ledger.Era (
  )
 import Cardano.Ledger.Mary (MaryEra)
 import Cardano.Ledger.Shelley.API (
+  CertState (..),
+  DState (..),
   EpochState (..),
   NewEpochState (..),
   StrictMaybe (..),
@@ -101,12 +104,49 @@ instance Crypto c => TranslateEra (AlonzoEra c) EpochState where
         , esNonMyopic = esNonMyopic es
         }
 
+instance Crypto c => TranslateEra (AlonzoEra c) DState where
+  translateEra _ ls =
+    pure
+      DState
+        { dsUnified = dsUnified ls
+        , dsIRewards = dsIRewards ls
+        , dsGenDelegs = dsGenDelegs ls
+        , dsFutureGenDelegs = dsFutureGenDelegs ls
+        }
+
+instance Crypto c => TranslateEra (AlonzoEra c) VState where
+  translateEra _ ls =
+    pure
+      VState
+        { vsDReps = vsDReps ls
+        , vsCCHotKeys = vsCCHotKeys ls
+        }
+
+instance Crypto c => TranslateEra (AlonzoEra c) PState where
+  translateEra _ ls =
+    pure
+      PState
+        { psStakePoolParams = psStakePoolParams ls
+        , psRetiring = psRetiring ls
+        , psFutureStakePoolParams = psFutureStakePoolParams ls
+        , psDeposits = psDeposits ls
+        }
+
+instance Crypto c => TranslateEra (AlonzoEra c) CertState where
+  translateEra ctxt ls =
+    pure
+      CertState
+        { certDState = translateEra' ctxt $ certDState ls
+        , certPState = translateEra' ctxt $ certPState ls
+        , certVState = translateEra' ctxt $ certVState ls
+        }
+
 instance Crypto c => TranslateEra (AlonzoEra c) API.LedgerState where
   translateEra ctxt ls =
     return
       API.LedgerState
         { API.lsUTxOState = translateEra' ctxt $ API.lsUTxOState ls
-        , API.lsDPState = API.lsDPState ls
+        , API.lsCertState = translateEra' ctxt $ API.lsCertState ls
         }
 
 instance Crypto c => TranslateEra (AlonzoEra c) API.UTxOState where