+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/LedgerState/NewEpochState.hs
@@ -18,15 +18,15 @@ import Cardano.Ledger.Address (isBootstrapRedeemer)
 import Cardano.Ledger.BaseTypes (
   BlocksMade (..),
  )
+import Cardano.Ledger.CertState (
+  CertState (..),
+  DState (..),
+  InstantaneousRewards (..),
+ )
 import Cardano.Ledger.Coin (
   Coin (..),
   CompactForm (CompactCoin),
   addDeltaCoin,
- )
-import Cardano.Ledger.DPState (
-  DPState (..),
-  DState (..),
-  InstantaneousRewards (..),
  )
 import Cardano.Ledger.Keys (
   GenDelegPair (..),
@@ -67,7 +67,7 @@ getGKeys nes = Map.keysSet genDelegs
   where
     NewEpochState _ _ _ es _ _ _ = nes
     EpochState _ _ ls _ _ _ = es
-    LedgerState _ (DPState (DState {dsGenDelegs = (GenDelegs genDelegs)}) _) = ls
+    LedgerState _ (CertState (DState {dsGenDelegs = (GenDelegs genDelegs)}) _ _) = ls
 
 -- | Creates the ledger state for an empty ledger which
 --  contains the specified transaction outputs.
@@ -86,7 +86,7 @@ genesisState genDelegs0 utxo0 =
         emptyGovernanceState
         (IStake mempty Map.empty)
     )
-    (DPState dState def)
+    (CertState dState def def)
   where
     dState = def {dsGenDelegs = GenDelegs genDelegs0}
 
@@ -106,8 +106,8 @@ depositPoolChange ls pp txBody = (currentPool <+> txDeposits) <-> txRefunds
     -- to emphasize this point.
 
     currentPool = (utxosDeposited . lsUTxOState) ls
-    txDeposits = totalTxDeposits pp (lsDPState ls) txBody
-    txRefunds = keyTxRefunds pp (lsDPState ls) txBody
+    txDeposits = totalTxDeposits pp (lsCertState ls) txBody
+    txRefunds = keyTxRefunds pp (lsCertState ls) txBody
 
 -- Remove the rewards from the UMap, but leave the deposits alone
 reapRewards ::