+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Core/Arbitrary.hs
@@ -52,11 +52,11 @@ import Cardano.Ledger.BaseTypes (
   textToUrl,
  )
 import Cardano.Ledger.Binary (EncCBOR, Sized, mkSized)
+import Cardano.Ledger.CertState
 import Cardano.Ledger.Coin (Coin (..), CompactForm (..), DeltaCoin (..))
 import Cardano.Ledger.Core
 import Cardano.Ledger.Credential (Credential (..), Ptr (..), StakeReference (..))
 import Cardano.Ledger.Crypto (Crypto (DSIGN))
-import Cardano.Ledger.DPState
 import Cardano.Ledger.EpochBoundary
 import Cardano.Ledger.Keys (
   GenDelegPair (..),
@@ -468,19 +468,23 @@ instance Crypto c => Arbitrary (UMap c) where
   shrink = genericShrink
 
 ------------------------------------------------------------------------------------------
--- Cardano.Ledger.DPState -------------------------------------------------------------------
+-- Cardano.Ledger.CertState -------------------------------------------------------------------
 ------------------------------------------------------------------------------------------
 
-instance Crypto c => Arbitrary (DPState c) where
-  arbitrary = DPState <$> arbitrary <*> arbitrary
+instance Era era => Arbitrary (CertState era) where
+  arbitrary = CertState <$> arbitrary <*> arbitrary <*> arbitrary
   shrink = genericShrink
 
-instance Crypto c => Arbitrary (DState c) where
+instance Era era => Arbitrary (DState era) where
   arbitrary = DState <$> arbitrary <*> arbitrary <*> arbitrary <*> arbitrary
   shrink = genericShrink
 
-instance Crypto c => Arbitrary (PState c) where
+instance Era era => Arbitrary (PState era) where
   arbitrary = PState <$> arbitrary <*> arbitrary <*> arbitrary <*> arbitrary
+  shrink = genericShrink
+
+instance Era era => Arbitrary (VState era) where
+  arbitrary = VState <$> arbitrary <*> arbitrary
   shrink = genericShrink
 
 instance Crypto c => Arbitrary (InstantaneousRewards c) where