+++ eras/mary/impl/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Version history for `cardano-ledger-mary`
 
+## 1.2.0.0
+
+* Replace `DPState c` with `CertState era`
+* Parametrize `DState` and `PState` by era
+* Add `TranslateEra` instances for:
+  * `DState`
+  * `PState`
+  * `VState`
+
 ## 1.1.0.0
 
 * Addition of `ToJSON` instances for `AssetName`, `PolicyID`, `MultiAsset` and `MaryValue`.